refactor: move installers from Install/ to Lang/
