+++ hie.yaml
@@ -33,19 +33,19 @@ cradle:
     - path: "site/app/Panbench/Shake/HTML.hs"
       component: "panbench-site:exe:site"
 
-    - path: "site/app/Panbench/Shake/Install/Agda.hs"
-      component: "panbench-site:exe:site"
-
-    - path: "site/app/Panbench/Shake/Install/Idris.hs"
-      component: "panbench-site:exe:site"
-
-    - path: "site/app/Panbench/Shake/Install/Lean.hs"
-      component: "panbench-site:exe:site"
-
-    - path: "site/app/Panbench/Shake/Install/Rocq.hs"
-      component: "panbench-site:exe:site"
-
     - path: "site/app/Panbench/Shake/Lang.hs"
+      component: "panbench-site:exe:site"
+
+    - path: "site/app/Panbench/Shake/Lang/Agda.hs"
+      component: "panbench-site:exe:site"
+
+    - path: "site/app/Panbench/Shake/Lang/Idris.hs"
+      component: "panbench-site:exe:site"
+
+    - path: "site/app/Panbench/Shake/Lang/Lean.hs"
+      component: "panbench-site:exe:site"
+
+    - path: "site/app/Panbench/Shake/Lang/Rocq.hs"
       component: "panbench-site:exe:site"
 
     - path: "site/app/Panbench/Shake/Make.hs"
